requirements.txt
amqp==5.3.1
anyio==4.9.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.3
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.1.7
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
kombu==5.5.3
lxml==5.4.0
oscrypto==1.3.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2==2.9.10
pycparser==2.22
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
pypdf==5.5.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-telegram-bot==22.0
PyYAML==6.0.2
qrcode==8.2
redis==5.2.1
reportlab==4.4.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.17
XlsxWriter==3.2.3
